Adding a cap on preview window.
The json string would get so long that the preview window would crash the whole page.
Adding a cap on it fixes this.
It will only display a preview for datasets smaller than 1001 entries. Any larger than that and a message saying that the preview is unavailable is shown instead.

changing message in preview window

Style fixes